move temp conversions to To2015B
